refactor(All) fix classes
