+++ internal/repositories/beju/repository.go
@@ -1,4 +1,4 @@
-package redis
+package beju
 
 import (
 	"context"
