+++ internal/repositories/platforms/pushover/pushhover.go
@@ -1,4 +1,4 @@
-package platforms
+package pushover
 
 import (
 	"fmt"
